+++ src/Symfony/Bundle/FrameworkBundle/Controller/AbstractController.php
@@ -410,7 +410,7 @@ abstract class AbstractController implements ServiceSubscriberInterface
     /**
      * Adds a Link HTTP header to the current response.
      *
-     * @see https://tools.ietf.org/html/rfc5988
+     * @see https://www.rfc-editor.org/rfc/rfc8288.html
      */
     protected function addLink(Request $request, LinkInterface $link): void
     {