+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/web_link.php
@@ -14,6 +14,10 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 use Symfony\Component\WebLink\EventListener\AddLinkHeaderListener;
 use Symfony\Component\WebLink\HttpHeaderParser;
 use Symfony\Component\WebLink\HttpHeaderSerializer;
+use Symfony\Component\WebLink\JsonLinksetParser;
+use Symfony\Component\WebLink\JsonLinksetSerializer;
+use Symfony\Component\WebLink\LinkTemplateHeaderParser;
+use Symfony\Component\WebLink\LinkTemplateHeaderSerializer;
 
 return static function (ContainerConfigurator $container) {
     $container->services()
@@ -24,9 +28,22 @@ return static function (ContainerConfigurator $container) {
         ->set('web_link.http_header_parser', HttpHeaderParser::class)
         ->alias(HttpHeaderParser::class, 'web_link.http_header_parser')
 
+        ->set('web_link.link_template_header_serializer', LinkTemplateHeaderSerializer::class)
+        ->alias(LinkTemplateHeaderSerializer::class, 'web_link.link_template_header_serializer')
+
+        ->set('web_link.link_template_header_parser', LinkTemplateHeaderParser::class)
+        ->alias(LinkTemplateHeaderParser::class, 'web_link.link_template_header_parser')
+
+        ->set('web_link.json_linkset_serializer', JsonLinksetSerializer::class)
+        ->alias(JsonLinksetSerializer::class, 'web_link.json_linkset_serializer')
+
+        ->set('web_link.json_linkset_parser', JsonLinksetParser::class)
+        ->alias(JsonLinksetParser::class, 'web_link.json_linkset_parser')
+
         ->set('web_link.add_link_header_listener', AddLinkHeaderListener::class)
             ->args([
                 service('web_link.http_header_serializer'),
+                service('web_link.link_template_header_serializer'),
             ])
             ->tag('kernel.event_subscriber')
     ;