+++ src/Symfony/Component/WebLink/Exception/ExceptionInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\WebLink\Exception;
+
+/**
+ * Base interface for all exceptions thrown by the WebLink component.
+ *
+ * @author Florent Morselli <florent.morselli@spomky-labs.com>
+ */
+interface ExceptionInterface extends \Throwable
+{
+}